% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.address}
\alias{chancer.address}
\title{Get a random address}
\usage{
chancer.address(abbreviation = FALSE)
}
\arguments{
\item{case}{can be 'upper' or 'lower'}

\item{len}{length of hash, defaults to 64}
}
\description{
Get a random address
}
\examples{
chancer.address()
chancer.address(abbreviation=TRUE)
}

