% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.age}
\alias{chancer.age}
\title{Get a random age}
\usage{
chancer.age(type = NULL)
}
\arguments{
\item{type}{if NULL defaults to age between 1-100 otherwise can be one of
'child', 'teen', 'adult', 'senior' to pick a pre-defined age range}
}
\description{
Get a random age
}
\examples{
chancer.age()
chancer.age(type='child')
chancer.age(type='senior')
}

