% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.altitude}
\alias{chancer.altitude}
\title{Get a random altitude}
\usage{
chancer.altitude(max = 8848)
}
\arguments{
\item{max}{maximum altitude in metres, defaults to 8848 (height of Everest)}
}
\description{
Get a random altitude
}
\examples{
chancer.altitude()
chancer.altitude(max=25)
}

