% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.areacode}
\alias{chancer.areacode}
\title{Get a random area code (US only)}
\usage{
chancer.areacode(parens = TRUE)
}
\arguments{
\item{parens}{TRUE or FALSE whether to wrap area code in parentheses}
}
\description{
Get a random area code (US only)
}
\examples{
chancer.areacode()
chancer.areacode(parens=FALSE)
}

