% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.birthday}
\alias{chancer.birthday}
\title{Get a random birthday}
\usage{
chancer.birthday(type = "POSIXlt")
}
\arguments{
\item{type}{eiter 'POSIXlt' to return a POSIXlt object or 'string' to return
date as a string}
}
\description{
Get a random birthday
}
\examples{
chancer.birthday()
chancer.birthday(type='string')
}

