% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.bool}
\alias{chancer.bool}
\title{Returns a random bool}
\usage{
chancer.bool(likelihood = 50)
}
\arguments{
\item{likelihood}{sets the likelihood of returning TRUE from 0-100}
}
\description{
Returns a random bool
}
\examples{
chancer.bool()
chancer.bool(10)
}

