% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.capitalise}
\alias{chancer.capitalise}
\title{Helper function to capitalise first char in string}
\usage{
chancer.capitalise(word)
}
\arguments{
\item{word}{string to capitalise}

\item{len}{length of hash, defaults to 64}
}
\description{
Helper function to capitalise first char in string
}
\examples{
chancer.hash()
chancer.hash(case='upper')
}

