% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.colour}
\alias{chancer.colour}
\title{Get a random colour}
\usage{
chancer.colour(format = "hex", greyscale = FALSE)
}
\arguments{
\item{format}{can be one of 'hex', 'rgb', 'shorthex'}

\item{greyscale}{TRUE or FALSE whether to restrict colour to greyscale}
}
\description{
Get a random colour
}
\examples{
chancer.colour()
chancer.colour(format='rgb')
chancer.colour(format='rgb', greyscale=TRUE)
}

