% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.coordinates}
\alias{chancer.coordinates}
\title{Get random coordinates}
\usage{
chancer.coordinates(format = "vector")
}
\arguments{
\item{format}{either 'vector' or 'string'}
}
\description{
Get random coordinates
}
\examples{
chancer.coordinates()
chancer.coordinates(format='string')
}

