% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.depth}
\alias{chancer.depth}
\title{Get a random depth (are always negative)}
\usage{
chancer.depth(min = -2250)
}
\arguments{
\item{min}{minimum value, defaults to -2550 (depth of the Mariana Trench)}
}
\description{
Get a random depth (are always negative)
}
\examples{
chancer.depth()
chancer.depth(min=-50)
}

