% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.domain}
\alias{chancer.domain}
\title{Get a random domain name}
\usage{
chancer.domain(tld = NULL)
}
\arguments{
\item{tld}{if NULL the tld is random or specifically set}
}
\description{
Get a random domain name
}
\examples{
chancer.domain()
chancer.domain(tld='ie')
}

