% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.email}
\alias{chancer.email}
\title{Get a random email address}
\usage{
chancer.email(domain = NULL)
}
\arguments{
\item{domain}{if NULL the domain is random or can be specifically set}
}
\description{
Get a random email address
}
\examples{
chancer.email()
chancer.email(domain='rmail.com')
}

