% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.first}
\alias{chancer.first}
\title{Get a random first name}
\usage{
chancer.first(gender = NULL)
}
\arguments{
\item{gender}{defaults to NULL but can get set to either 'male' or 'female'}
}
\description{
Get a random first name
}
\examples{
chancer.first()
chancer.first(gender='female')
}

