% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.floating}
\alias{chancer.floating}
\title{Get a random floating point number}
\usage{
chancer.floating(min = 0, max = 100, fixed = 4)
}
\arguments{
\item{min}{minimum possible value}

\item{max}{maximum possible value}

\item{fixed}{maximum number of decimal place to allow. Note may return
fewer decimal places, fixed is not guaranteed}
}
\description{
Get a random floating point number
}
\examples{
chancer.floating(0, 1, 2)
}

