% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.geohash}
\alias{chancer.geohash}
\title{Get a random geohash}
\usage{
chancer.geohash(len = 7)
}
\arguments{
\item{len}{numbers of characters of accuracy (defaults to 7)}
}
\description{
Get a random geohash
}
\examples{
chancer.geohash()
chancer.geohash(len=3)
}

