% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.hash}
\alias{chancer.hash}
\title{Get a random hex hash}
\usage{
chancer.hash(len = 64, case = "lower")
}
\arguments{
\item{len}{length of hash, defaults to 64}

\item{case}{can be 'upper' or 'lower'}
}
\description{
Get a random hex hash
}
\examples{
chancer.hash()
chancer.hash(case='upper')
}

