% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.integer}
\alias{chancer.integer}
\title{Get a random integer}
\usage{
chancer.integer(min = MIN_INT, max = MAX_INT)
}
\arguments{
\item{min}{minimum possible value}

\item{max}{maximum possible value}
}
\description{
Get a random integer
}
\examples{
chancer.integer()
chancer.integer(1, 25)
}

