% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.name}
\alias{chancer.name}
\title{Get a random name}
\usage{
chancer.name(initial = FALSE, middle_name = FALSE, gender = NULL,
  prefix = FALSE)
}
\arguments{
\item{initial}{TRUE or FALSE to include a middle initial}

\item{middle_name}{TRUE or FALSE to include a middle name}

\item{gender}{if NULL gender is randomly selected else can be one of
'male' or 'female'}

\item{prefix}{TRUE or FALSE to include a prefix to the name, e.g Doctore, Mister etc}
}
\description{
Get a random name
}
\examples{
chancer.name()
chancer.name(initial=TRUE)
chancer.name(middle_name=TRUE)
chancer.name(middle_name=TRUE, prefix=TRUE)
}

