% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.natural}
\alias{chancer.natural}
\title{Get a random natural number}
\usage{
chancer.natural(min = 1, max = 100)
}
\arguments{
\item{min}{minimum possible value}

\item{max}{maximum possible value}
}
\description{
Get a random natural number
}
\examples{
chancer.natural()
chancer.natural(1, 25)
}

