% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.paragraph}
\alias{chancer.paragraph}
\title{Get a random sentence}
\usage{
chancer.paragraph(sentences = NULL)
}
\arguments{
\item{sentences}{optional to set maximum sentences in paragraph otherwise defaults to
random length between 3 - 7 sentences}
}
\description{
Get a random sentence
}
\examples{
chancer.paragraph()
chancer.paragraph(sentences=4)
}

