% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.pick}
\alias{chancer.pick}
\title{Select a random element from a vector}
\usage{
chancer.pick(vec = NULL)
}
\arguments{
\item{vec}{vector to select from}
}
\description{
Select a random element from a vector
}
\examples{
chancer.pick(c(1, 2, 3, 4))
}

