% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.prefix}
\alias{chancer.prefix}
\title{Get a random prefix for name, e.g Mister, Doctor etc}
\usage{
chancer.prefix(gender = NULL, abbreviated = FALSE)
}
\arguments{
\item{gender}{one of 'male', 'female', 'all'}

\item{abbreviated}{boolean whether to abbreviate the prefix or not}
}
\description{
Get a random prefix for name, e.g Mister, Doctor etc
}
\examples{
chancer.prefix()
chancer.prefix('male', TRUE)
}

