% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.sentence}
\alias{chancer.sentence}
\title{Get a random sentence}
\usage{
chancer.sentence(words = NULL)
}
\arguments{
\item{words}{optional to set maximum words in sentence otherwise defaults to
random length between 12 - 18 words}
}
\description{
Get a random sentence
}
\examples{
chancer.sentence()
chancer.sentence(words=4)
}

