% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.street}
\alias{chancer.street}
\title{Get a random street name}
\usage{
chancer.street(abbreviation = FALSE)
}
\arguments{
\item{abbreviation}{TRUE or FALSE whether to abbreviate street name,
e.g Street to St}
}
\description{
Get a random street name
}
\examples{
chancer.street()
chancer.street(abbreviation=TRUE)
}

