% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.string}
\alias{chancer.string}
\title{Get a random string}
\usage{
chancer.string(len = 10, case = "all", pool = "all", user = NULL)
}
\arguments{
\item{len}{length of string to return}

\item{case}{select 'uppercase', 'lowercase' or 'all' for both cases}

\item{pool}{select 'alpha' for just letters, 'numeric' for just numbers,
'symbol' for just symbols or 'all' for all}

\item{user}{if null string contains chars from
'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*()'
else you can pass in a user defined string of chars}
}
\description{
Get a random string
}
\examples{
chancer.string(len=10, user='xyz')
chancer.string(len=10, case='lower', pool='alpha')
chancer.string(len=10, pool='symbol')
}

