% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.weighted}
\alias{chancer.weighted}
\title{Selects a random element from a vector using weighted probabilities}
\usage{
chancer.weighted(vec = NULL, probs = NULL)
}
\arguments{
\item{vec}{vector to select from}
}
\description{
Selects a random element from a vector using weighted probabilities
}
\examples{
chancer.weighted(c(1, 2, 3), c(0.5, 0.4, 0.1))
}

