% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chancer.word}
\alias{chancer.word}
\title{Get a random word}
\usage{
chancer.word(syllables = NULL, len = NULL)
}
\arguments{
\item{syllables}{optional maximum number of syllables to use. Note that only
one of len and syllables can be used}

\item{len}{optional maximum length of string to return. Note that only
one of len and syllables can be used}
}
\description{
Get a random word
}
\examples{
chancer.word()
chancer.string(len=10, case='lower', pool='alpha')
chancer.string(len=10, pool='symbol')
}

